RangeTree change value Interface
Add RangeTree constructor with

+++ DebugConsole/RangeTreeData.cs
@@ -18,8 +18,8 @@ namespace DebugConsole
             _y = y;
         }
 
-        public double X => _x;
+        public float X => _x;
 
-        public double Y => _y;
+        public float Y => _y;
     }
 }
